Flytter ut kode i egen komponent
