+++ src/frontend/Komponenter/Infotrygd/typer.ts
@@ -1,0 +1,5 @@
+import { InfotrygdPeriode, Kode } from '../../App/typer/infotrygd';
+
+export type InfotrygdPeriodeMedFlereEndringer = InfotrygdPeriode & {
+    initialKode?: Kode;
+};
